requirements.txt
# ── requirements.txt ─────────────────────────────────────────────────────
# AI Trading Assistant — All Dependencies
# Install with: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────


# ── Web UI ────────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Data Processing ───────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Machine Learning ──────────────────────────────────────────────────────
scikit-learn>=1.3.0

# ── Charts & Visualization ────────────────────────────────────────────────
plotly>=5.18.0

# ── Groq AI (LLaMA3) ─────────────────────────────────────────────────────
groq>=0.5.0

# ── Environment Variables ─────────────────────────────────────────────────
python-dotenv>=1.0.0